Split onboarding action handlers to satisfy the 800-line file-size gate.
